refactor: ユーザ API のリファクタリング
- UseCase の作成
- FormRequest のクラス名変更等・ディレクトリ移動
- 返り値を JsonResource に変更
