+++ app/UseCases/User/UpdateUser.php
@@ -1,0 +1,15 @@
+<?php
+
+namespace App\UseCases\User;
+
+use App\Models\User;
+
+class UpdateUser
+{
+    public function __invoke(User $user): User
+    {
+        $user->save();
+
+        return $user;
+    }
+}
